BUilding a more complete set of tools to monitor performances
