+++ libraries/render/src/render/Engine.h
@@ -13,7 +13,6 @@
 #define hifi_render_Engine_h
 
 #include <SettingHandle.h>
-#include <gpu/Context.h>
 
 #include "Context.h"
 #include "Task.h"
@@ -48,57 +47,6 @@ namespace render {
         RenderContextPointer _renderContext;
     };
     using EnginePointer = std::shared_ptr<Engine>;
-
-
-    // A simple job collecting global stats on the Engine / Scene / GPU
-    class EngineStatsConfig : public Job::Config{
-        Q_OBJECT
-
-        Q_PROPERTY(int numBuffers MEMBER numBuffers NOTIFY dirty)
-        Q_PROPERTY(int numGPUBuffers MEMBER numGPUBuffers NOTIFY dirty)
-        Q_PROPERTY(qint64 bufferSysmemUsage MEMBER bufferSysmemUsage NOTIFY dirty)
-        Q_PROPERTY(qint64 bufferVidmemUsage MEMBER bufferVidmemUsage NOTIFY dirty)
-
-        Q_PROPERTY(int numTextures MEMBER numTextures NOTIFY dirty)
-        Q_PROPERTY(int numGPUTextures MEMBER numGPUTextures NOTIFY dirty)
-        Q_PROPERTY(qint64 textureSysmemUsage MEMBER textureSysmemUsage NOTIFY dirty)
-        Q_PROPERTY(qint64 textureVidmemUsage MEMBER textureVidmemUsage NOTIFY dirty)
-        Q_PROPERTY(int numFrameTextures MEMBER numFrameTextures NOTIFY dirty)
-    public:
-        EngineStatsConfig() : Job::Config(true) {}
-
-        int numBuffers{ 0 };
-        int numGPUBuffers{ 0 };
-        qint64 bufferSysmemUsage{ 0 };
-        qint64 bufferVidmemUsage{ 0 };
-
-        int numTextures{ 0 };
-        int numGPUTextures{ 0 };
-        qint64 textureSysmemUsage{ 0 };
-        qint64 textureVidmemUsage{ 0 };
-
-        int numFrameTriangles{ 0 };
-        int numFrameTextures{ 0 };
-
-        void emitDirty() { emit dirty(); }
-
-    signals:
-        void dirty();
-    };
-
-    class EngineStats {
-    public:
-        using Config = EngineStatsConfig;
-        using JobModel = Job::Model<EngineStats, Config>;
-
-        EngineStats() {}
-
-        gpu::ContextStats _gpuStats;
-
-        void configure(const Config& configuration) {}
-        void run(const SceneContextPointer& sceneContext, const RenderContextPointer& renderContext);
-    };
-
 
 }
